Renaming DB services.
